requirements.txt
# ctransformers==0.2.5
# sentence-transformers==2.2.2
# llama-cpp-python==0.2.54
# python-dotenv==1.0.0
# tqdm==4.65.0
# langchain==0.0.352
# chromadb==0.3.23

langchain==0.0.171
pygpt4all==1.1.0
chromadb==0.3.23
llama-cpp-python==0.2.55
urllib3==2.0.2
pdfminer.six==20221105
python-dotenv==1.0.0
unstructured==0.6.6
extract-msg==0.41.1
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.65.0